requirements.txt
mysql-connector-python
pip install -r requirements.txt